[PC-1268] Fixes following code review
