+++ src/components/hocs/with-login/__specs__/withRequiredLogin.spec.js
@@ -20,7 +20,11 @@ const RequiredLoginTest = withRequiredLogin(Test)
 
 jest.mock('../helpers')
 
-describe('src | components | pages | hocs | with-login | withRequiredLogin', () => {
+describe('src | components | pages | hocs | with-login | withRequiredLogin - unit tests', () => {
+  beforeEach(() => {
+    fetch.resetMocks()
+  })
+
   describe('snapshot', () => {
     it('should match snapshot', () => {
       // when
@@ -32,106 +36,8 @@ describe('src | components | pages | hocs | with-login | withRequiredLogin', () 
     })
   })
 
-  beforeEach(() => {
-    fetch.resetMocks()
-  })
-
-  describe('functions', () => {
-    it('should redirect to signin when not authenticated', () => {
-      return new Promise(done => {
-        // given
-        const history = createBrowserHistory()
-        history.push('/test')
-        const store = configureTestStore()
-        fetch.mockResponse(JSON.stringify([{ global: ['Nobody is authenticated here'] }]), {
-          status: 400,
-        })
-
-        // when
-        mount(
-          <Provider store={store}>
-            <Router history={history}>
-              <Switch>
-                <Route path="/test">
-                  <RequiredLoginTest />
-                </Route>
-                <Route path="/connexion">
-                  <OnMountCaller onMountCallback={done} />
-                </Route>
-              </Switch>
-            </Router>
-          </Provider>
-        )
-      })
-    })
-
-    it('should redirect to typeform when authenticated and not needsToFillCulturalSurvey', done => {
-      // given
-      const history = createBrowserHistory()
-      history.push('/test')
-      const store = configureTestStore()
-      fetch.mockResponse(
-        JSON.stringify({
-          email: 'michel.marx@youpi.fr',
-          needsToFillCulturalSurvey: true,
-        }),
-        {
-          status: 200,
-        }
-      )
-
-        // when
-        mount(
-          <Provider store={store}>
-            <Router history={history}>
-              <Switch>
-                <Route path="/test">
-                  <RequiredLoginTest />
-                </Route>
-                <Route path="/typeform">
-                  <OnMountCaller onMountCallback={done} />
-                </Route>
-              </Switch>
-            </Router>
-          </Provider>
-        )
-      })
-    })
-
-    it('should not redirect when authenticated and needsToFillCulturalSurvey', () => {
-      // given
-      const history = createBrowserHistory()
-      history.push('/test')
-      const store = configureTestStore()
-      fetch.mockResponse(
-        JSON.stringify({
-          email: 'michel.marx@youpi.fr',
-          needsToFillCulturalSurvey: false,
-        }),
-        {
-          status: 200,
-        }
-      )
-
-      // when
-      mount(
-        <Provider store={store}>
-          <Router history={history}>
-            <Switch>
-              <Route path="/test">
-                <RequiredLoginTest />
-              </Route>
-            </Switch>
-          </Router>
-        </Provider>
-      )
-
-      expect(history.location.pathname).toBe('/test')
-    })
-  })
-
   describe('handleFail()', () => {
-    it('should call push history when its fail', () => {
+    it('should call push history when it fails', () => {
       // given
       const state = {}
       const action = {}
@@ -156,13 +62,25 @@ describe('src | components | pages | hocs | with-login | withRequiredLogin', () 
       // given
       const state = {}
       const action = {
-        payload: {},
+        payload: {
+          datum: {
+            email: 'michel.marx@youpi.fr',
+            needsToFillCulturalSurvey: false,
+          }
+        }
       }
+
       const ownProps = {
         history: {
           push: jest.fn(),
         },
-        location: {},
+        location: {
+          hash: '',
+          key: expect.any(String),
+          pathname: '/test',
+          search: '',
+          state: undefined,
+        },
       }
 
       // when
@@ -182,7 +100,7 @@ describe('src | components | pages | hocs | with-login | withRequiredLogin', () 
       })
     })
 
-    it('should not call push history when user is redirected', () => {
+    it('should call push history when user is redirected', () => {
       // given
       const state = {}
       const action = {
@@ -203,7 +121,7 @@ describe('src | components | pages | hocs | with-login | withRequiredLogin', () 
       expect(ownProps.history.push).toHaveBeenCalledWith('/fake-url')
     })
 
-    it('should call push history when its success', () => {
+    it('should not call push history when successful', () => {
       // given
       const state = {}
       const action = {
@@ -221,7 +139,7 @@ describe('src | components | pages | hocs | with-login | withRequiredLogin', () 
       handleSuccess(state, action, ownProps)
 
       // then
-      expect(ownProps.history.push).not.toBeCalled()
+      expect(ownProps.history.push).not.toHaveBeenCalled()
     })
   })
 })
