+++ src/components/pages/discovery/specs/DeckNavigation.spec.js
