+++ src/components/pages/discovery/specs/Discovery.spec.js
