+++ src/components/pages/discovery/specs/DiscoveryContainer.spec.js
